% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mobloc.R
\name{calculate_mobloc}
\alias{calculate_mobloc}
\title{Calculate the estimated location of mobile devices}
\usage{
calculate_mobloc(prop, prior, raster)
}
\arguments{
\item{prop}{a propagation object, which is the result of \code{\link{process_cellplan}}}

\item{prior}{prior object, the result of \code{\link{create_uniform_prior}}, \code{\link{create_prior}}, or \code{\link{create_network_prior}}}

\item{raster}{raster object that contains the raster tile index numbers (e.g. created with \code{\link{create_raster}})}
}
\description{
Calculate the estimated location of mobile devices, i.e. the location posterior probabilities. This is the final step of the mobloc process (see \href{../doc/mobloc.html}{\code{vignette("mobloc")}}).
}
\seealso{
\href{../doc/mobloc.html}{\code{vignette("mobloc")}}
}
