% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_inspection_tool.R
\name{cell_inspection_tool}
\alias{cell_inspection_tool}
\title{Visualize signal strengths and probabilities per raster cell}
\usage{
cell_inspection_tool(cp, cp_poly, raster, prob, param, prior = NULL,
  tm = NULL)
}
\arguments{
\item{cp}{cellplan}

\item{cp_poly}{cellplan polygons}

\item{raster}{raster}

\item{prob}{probabilities per raster cell}

\item{param}{parameter list}
}
\description{
Visualize signal strengths and probabilities per raster cell
}
