% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_modelling_tool.R
\name{cell_modelling_tool}
\alias{cell_modelling_tool}
\title{Interactive cell modelling tool}
\usage{
cell_modelling_tool(param, plot.height = 800)
}
\arguments{
\item{param}{list of parameters}

\item{plot.height}{heigth of the plots in the tool (in pixels)}
}
\description{
Tool to model the perceived signal strength
}
