% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cellplan.R
\name{check_cellplan}
\alias{check_cellplan}
\title{Check cellplan}
\usage{
check_cellplan(cp, param, elevation = NULL)
}
\arguments{
\item{cp}{cellplan \code{sf} object containing the antenna data. Each data record should be a point (i.e., `st_geometry_type(cp)` should return `POINT`s). The variables (of which only the first is required) are used:
\itemize{
\item \code{z} or \code{height} (required). Both indicate the height of the antenna, but \code{z} is including and \code{height} is excluding the elevation. If only \code{height} is present, also the argument \code{elevation} is required.
\item \code{direction}. Direction of the antanna in degrees. Use \code{NA} for omnidirectional antennas.
\item \code{tilt}. Tilt of the antennas in degrees. Only applicable for directional cells. If omitted, the default value \code{tilt} from the parameter list \code{param} will be used.
\item \code{beam_h}. Horizontal beam width in degrees. The signal loss at \code{-beam_h/2} and \code{+beam_h/2} degrees is 3 dB. Run \code{radiation_plot(beam_width = 65, db_back = -30)}. If omitted, the default value \code{beam_h} from the parameter list \code{param} will be used.
\item \code{beam_v}. Vertical beam width in degrees. The signal loss at \code{-beam_v/2} and \code{+beam_v/2} degrees is 3 dB. Run \code{radiation_plot(type = "e", beam_width = 9, db_back = -30)}. If omitted, the default value \code{beam_v} from the parameter list \code{param} will be used.
\item \code{small}. Logical value that determines whether the antenna is a 'small cell'.
\item \code{range}. The maximum range of the antenna. If omitted, the value \code{max_range} from the parameter list \code{param} will be used. If \code{small} is defined, the value \code{max_range_small} is used for each antenna for which \code{small == TRUE}.
}
Required variables:}
}
\description{
Function to the check cellplan, i.e. the format and whether are all required variables present.
}
