% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{combine_raster_layers}
\alias{combine_raster_layers}
\title{Combine raster layers to a single raster layer}
\usage{
combine_raster_layers(..., weights = NULL)
}
\arguments{
\item{...}{Either one raster brick/stack, or multiple raster layer objects}

\item{weights}{numeric vector that specifies the weights of the raster layers}
}
\value{
raster object of one layer
}
\description{
Combine raster layers to a single raster layer. Weights can be assigned to the raster layers.
}
\examples{
# load data
data("ZL_landuse")

# create environment layer
ZL_envir <- combine_raster_layers(ZL_landuse, weights = c(1, 1, 1, 0, 0))

if (requireN(tmap)) {
    tmap_mode("view")
    qtm(ZL_envir)
}
}
\seealso{
\code{\link{create_prior}}
}
