% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{composite_prior}
\alias{composite_prior}
\title{Create a composite prior}
\usage{
composite_prior(priorlist, composition = NULL)
}
\arguments{
\item{priorlist}{list of priors (see \code{\link{extract_prior}})}

\item{composition}{composition of priors, a vector with fractions that should add up to 1}
}
\value{
data frame with two columns, raster id number \code{rid} and probability \code{p}. These probabilities will add up to 1.
}
\description{
Create a composite prior
}
