% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sig_strength.R
\name{compute_sig_strength}
\alias{compute_sig_strength}
\title{Compute signal strength}
\usage{
compute_sig_strength(cp, raster, elevation, param, region = NULL)
}
\arguments{
\item{cp}{cellplan, validated with \code{\link{validate_cellplan}}}

\item{raster}{raster object that contains the raster tile index numbers (e.g. created with \code{\link{create_raster}})}

\item{elevation}{raster with elevation data}

\item{param}{parameter list created with \code{mobloc_param}}

\item{region}{polygon shape. If specified, only the signal strength will be calculated for raster tiles inside the polygons}
}
\value{
a data.frame is return with the following colums: cell (cell id), rid (raster tile id), dist (distance between cell and grid tile), dBm (signal strength), s (signal dominance), pag (likelihood probability). This data.frame is required to run the interactive tool \code{explore_mobloc} from the \code{mobvis} package and to compute the connection likelihood with \code{\link{create_strength_llh}}.
}
\description{
Compute signal strength (propagation). The propagation is modelled based on the physical properties of the cells. Also, the likelihood distribution is calculated, which takes the overlap of cells into account.
}
\examples{
\dontrun{
# set parameters
ZL_param <- mobloc_param()

# load data
data("ZL_cellplan", "ZL_muni", "ZL_elevation", "ZL_landuse")

# create environment layer (needed to calculate path loss exponent (ple))
ZL_envir <- combine_raster_layers(ZL_landuse, weights = c(1, 1, 1, 0, 0))

# validate cellplan
ZL_cellplan <- validate_cellplan(ZL_cellplan, param = ZL_param, region = ZL_muni,
                                 envir = ZL_envir, elevation = ZL_elevation)

# create raster
ZL_bbox <- sf::st_bbox(c(xmin = 4012000, ymin = 3077000, xmax = 4048000, ymax = 3117000),
                       crs = sf::st_crs(3035))
ZL_raster <- create_raster(ZL_bbox)

# compute the signal strength model
ZL_strength <- compute_sig_strength(cp = ZL_cellplan, raster = ZL_raster,
                                    elevation = ZL_elevation, param = ZL_param)
}
}
\seealso{
\href{../doc/mobloc.html}{\code{vignette("mobloc")}}
}
