% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bbx.R
\name{create_bbx_rect}
\alias{create_bbx_rect}
\title{Create a bounding box rectangle}
\usage{
create_bbx_rect(bbx, stepsize = 10)
}
\arguments{
\item{bbx}{bounding box}

\item{stepsize}{stepsize}
}
\description{
Create a bounding box rectangle. Code borrowed from \code{tmaptools:::create_sf_rect}
}
