% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{create_prior}
\alias{create_prior}
\alias{create_uniform_prior}
\alias{create_network_prior}
\alias{prior_filter}
\title{Create prior}
\usage{
create_prior(..., name = "composite", weights = NULL)

create_uniform_prior(raster)

create_network_prior(prop, raster)

prior_filter(prior, region)
}
\arguments{
\item{...}{one of more raster objects, or a list of it. When multiple raster layers are used, these are first combined using a weighed sum (see argument \code{weights} and the underlying function \code{\link{combine_raster_layers}}). After that, these values are normalised such that the values add up to 1}

\item{name}{name of the prior}

\item{weights}{The weights of the raster objects (see \code{...}). See also the underlying function \code{\link{combine_raster_layers}}.}

\item{raster}{a raster object that contains the raster numbers (e.g. created with \code{\link{create_raster}})}

\item{prop}{a propagation object, which is the result of \code{\link{process_cellplan}}}

\item{prior}{object created with \code{create_prior}, \code{create_network_prior} or \code{create_uniform_prior}}

\item{region}{a polygons object (class \code{sf}) that specifies a region}
}
\value{
data frame with two columns, raster id number \code{rid} and probability \code{p}. These probabilities will add up to 1.
}
\description{
Functions to create a prior. A prior is a raster layer where all values add up to 1. The function \code{create_prior} creates a prior using one or more raster layers. The function \code{create_uniform_prior} extracts a uniform prior from the propagation model. The function \code{create_network_prior} extracts a prior from the propagation model. The function \code{prior_filter} can be used to filter a prior raster using a polygons object.
}
