% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_rasterize.R
\name{create_raster}
\alias{create_raster}
\title{Create empty raster}
\usage{
create_raster(bbx, cell.size = 100)
}
\arguments{
\item{bbx}{bounding box (see \code{\link[sf:st_bbox]{st_bbox}}), where the projection (\code{crs}) must be specified}

\item{cell.size}{cell size (which is both the width and the height) in meters (assuming the coordinate system is specified in meters)}
}
\value{
raster layer
}
\description{
Create empty raster from a bounding box with a specified projection.
}
