% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_strength_llh.R
\name{create_strength_llh}
\alias{create_strength_llh}
\title{Create signal strenth likelihood}
\usage{
create_strength_llh(strength, param)
}
\arguments{
\item{strength}{a signal strength model object, which is the result of \code{\link{compute_sig_strength}}}

\item{param}{parameter list created with \code{mobloc_param}}
}
\description{
Create signal strenth likelihood
}
