% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_voronoi_llh.R
\name{create_voronoi_llh}
\alias{create_voronoi_llh}
\title{Create Voronoi likelihood}
\usage{
create_voronoi_llh(cp, raster, offset = 100)
}
\arguments{
\item{cp}{cellplan, validated with \code{\link{validate_cellplan}}}

\item{raster}{raster object that contains the raster tile index numbers (e.g. created with \code{\link{create_raster}})}

\item{offset}{offset in meters of the cells in the direction of progagation}
}
\description{
Create Voronoi likelihood
}
