% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cluster.R
\name{current_cluster}
\alias{current_cluster}
\title{Returns number of cors in the current cluster}
\usage{
current_cluster(verbose = TRUE)
}
\arguments{
\item{verbose}{should warnings be given?}
}
\value{
number of cores
}
\description{
Returns number of cors in the current cluster
}
