% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_strength.R
\name{dBW2dBm}
\alias{dBW2dBm}
\alias{dBW2W}
\alias{W2dBW}
\alias{W2dBm}
\alias{dBm2dBW}
\alias{dBm2W}
\title{Converter functions}
\usage{
dBW2dBm(dBW)

dBW2W(dBW)

W2dBW(W)

W2dBm(W)

dBm2dBW(dBm)

dBm2W(dBm)
}
\arguments{
\item{dBW}{decibelwatt}

\item{W}{watt}

\item{dBm}{decibel-milliwatts, a unit of level used to indicate that a power ratio is expressed in decibels (dB) with reference to one milliwatt (mW)}
}
\description{
Converter functions between W, dBW, and dBm.
}
