% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_antennas.R
\name{explore_antennas}
\alias{explore_antennas}
\title{Explore the Visualize signal strengths and probabilities per raster cell}
\usage{
explore_antennas(cp, cp_poly, raster, propagation, priorlist = NULL,
  param, tm = NULL)
}
\arguments{
\item{cp}{cellplan}

\item{cp_poly}{cellplan polygons}

\item{raster}{raster}

\item{propagation}{probabilities per raster cell}

\item{param}{parameter list}
}
\description{
Visualize signal strengths and probabilities per raster cell
}
