% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_propagation_model.R
\name{explore_propagation_model}
\alias{explore_propagation_model}
\title{Interactive exploration tool for the propagation model}
\usage{
explore_propagation_model(param, plot.height = 800)
}
\arguments{
\item{param}{list of parameters}

\item{plot.height}{heigth of the plots in the tool (in pixels)}
}
\description{
Tool to setup the propagation model parameters
}
