% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{create_uniform_prior}
\alias{create_uniform_prior}
\alias{create_network_prior}
\alias{extract_prior}
\alias{create_prior}
\title{Extract prior information}
\usage{
create_uniform_prior(raster)

create_network_prior(prop, raster)

create_prior(..., name = "composite", weights = NULL)
}
\arguments{
\item{prop}{a propagation object, which is the result of \code{\link{process_cellplan}}}

\item{x}{a raster object with per raster cell a prior value.}
}
\value{
data frame with two columns, raster id number \code{rid} and probability \code{p}. These probabilities will add up to 1.
}
\description{
Extract prior information. The function \code{extract_prior} extracts the prior from a raster object. The function \code{extract_uniform_prior} extracts a uniform prior from the propagation model. The function \code{extract_network_prior} extracts a prior from the propagation model.
}
