% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_model_parameters.R
\name{location_model_parameters}
\alias{location_model_parameters}
\alias{update_model_parameters}
\title{Set and update location model parameters}
\usage{
location_model_parameters(db0_tower = -45, db0_small = -60,
  delta_min3dB = 65, delta_dB_back = -30, epsilon_min3dB = 9,
  epsilon_dB_back = -30, db_mid = -92.5, db_width = 5,
  poly_shape = "plectrum", max_range = 10000, max_range_small = 100,
  area_expension = 4, max_overlapping_cells = 20)

update_model_parameters(param, ...)
}
\arguments{
\item{db0_tower}{signal strength in dBm near a large cell (placed in a cell tower or rooftop site)}

\item{db0_small}{signal strength in dBm near a small cell (omnidirectional)}

\item{delta_min3dB}{default horizontal beam width. At \code{delta_min3dB/2}, the signal strength is halved (so -3dBM)}

\item{delta_dB_back}{difference in signal strength between front and back}

\item{epsilon_min3dB}{default vertical beam width. At \code{epsilon_min3dB/2}, the signal strength is halved (so -3dBM)}

\item{epsilon_dB_back}{difference in signal strength between front and back}

\item{db_mid}{middle point in the logistic function to map signal strength to probability}

\item{db_width}{width of the logistic function to map signal strength to probability}

\item{poly_shape}{shape of the polygon that defines the coverage area of a cell. One of \code{"pie"}, \code{"plectrum"}, \code{"Voronoi"}.}

\item{max_range}{maximum range of large cells}

\item{max_range_small}{maximum range of small cells}

\item{area_expension}{when \code{poly_shape} is \code{"pie"} or \code{"plectrum"}, the size is determined by the corresponding Voronoi size and multiplied by \code{area_expension} to allow overlap.}

\item{max_overlapping_cells}{maximum number of polygons that may overlap per raster cell. If the actual number exceeds this parameter, the \code{max_overlapping_cells} cells with the highest signal strength are selected}

\item{param}{parameter list created with \code{location_model_parameters}}

\item{...}{parameter updates, see arguments of \code{location_model_parameters}}
}
\value{
parameter list
}
\description{
Set and update location model parameters. The function \code{location_model_parameters} specifies model parameters, which can be updated with \code{update_model_parameters}
}
