% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{mobloc-package}
\alias{mobloc-package}
\alias{mobloc}
\title{Mobile location algorithms and tools}
\description{
CDR and signalling data contain mobile phone events logged at specific cells. This package provides algorithms and tools to translate cell properties to gespatial distributions. Depending on the availability of the cell properties, such as coordinates, direction, and height, different algorithms can be used.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{Mobile}
\keyword{Voronoi}
\keyword{data,}
\keyword{location,}
\keyword{phone}
