% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_cells_into_prop_direction.R
\name{move_cells_into_prop_direction}
\alias{move_cells_into_prop_direction}
\title{Move cells into propagation direction}
\usage{
move_cells_into_prop_direction(cp, offset = 100)
}
\arguments{
\item{cp}{cellplan, validated with \code{\link{validate_cellplan}}}

\item{offset}{offset in meters}
}
\description{
Move cells into propagation direction
}
