% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_strength.R
\name{norm_dBloss}
\alias{norm_dBloss}
\alias{distance2dB}
\alias{db2s}
\title{Signal strength model functions}
\usage{
norm_dBloss(a, db_back = -30, sd = 90, beam_width = NULL)

distance2dB(r, ple, W)

db2s(dBm, midpoint, steepness)
}
\arguments{
\item{a}{angle of direction (in degrees)}

\item{db_back}{difference in signal strength between front and back (in dB)}

\item{sd}{standard deviation of the normal distribution that is used to model the radiation pattern, such that at 180 degrees, the dB loss is db_back and at beam_width degrees the dB loss is 3dB.}

\item{beam_width}{beam width}

\item{r}{distance}

\item{ple}{path loss exponent}

\item{W}{watt}

\item{dBm}{dBm to convert}

\item{midpoint}{which dBm value corresponds to the mid point of the distribution (0.5)}

\item{steepness}{steepness of the fitted curve}
}
\description{
Signal strength model functions. \code{norm_dBloss} approximates the dB loss at a certain direction using normal distributions. \code{distance2dB} approximates the signal strength in dB as a function of distance, given the power of the cell and the path loss exponent. \code{db2s} transforms dBm to signal dominance (s) using a sigmoid function.
}
