% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cellplan.R
\name{process_cellplan}
\alias{process_cellplan}
\title{Rasterize cellplan}
\usage{
process_cellplan(cp, raster, elevation, param, region = NULL)
}
\arguments{
\item{cp}{cellplan}

\item{raster}{raster with indices}

\item{elevation}{raster with elevation data}

\item{param}{list}

\item{region}{polygon shape. If specified, only the signal strength will be calculated for raster cells inside the polygons}
}
\description{
Rasterize cellplan
}
