% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation_parameters.R
\name{propagation_parameters}
\alias{propagation_parameters}
\title{Set and update location model parameters}
\usage{
propagation_parameters(db0_tower = -45, db0_small = -60,
  azim_min3dB = 65, azim_dB_back = -30, elev_min3dB = 9,
  elev_dB_back = -30, db_mid = -92.5, db_width = 5,
  poly_shape = "oval", max_range = 10000, max_range_small = 100,
  min_range = 400, min_range_small = 100, height = 30,
  height_small = 8, tilt = 5, beam_v = 9, beam_h = 65,
  area_expension = 4, max_overlapping_cells = 20)
}
\arguments{
\item{db0_tower}{signal strength in dBm near a normal antennas (placed in a cell tower or rooftop site)}

\item{db0_small}{signal strength in dBm near a small cell (omnidirectional)}

\item{azim_min3dB}{default horizontal beam width. At \code{azim_min3dB/2}, the signal strength is halved (so -3dBM)}

\item{azim_dB_back}{difference in signal strength between front and back}

\item{elev_min3dB}{default vertical beam width. At \code{elev_min3dB/2}, the signal strength is halved (so -3dBM)}

\item{elev_dB_back}{difference in signal strength between front and back}

\item{db_mid}{middle point in the logistic function to map signal strength to probability}

\item{db_width}{width of the logistic function to map signal strength to probability}

\item{poly_shape}{shape of the polygon that defines the coverage area of a cell. One of \code{"pie"}, \code{"oval"} (default), \code{"Voronoi"}.}

\item{max_range}{maximum range of normal antennas}

\item{max_range_small}{maximum range of small cells}

\item{min_range}{minimum range of normal antennas}

\item{min_range_small}{minimum range of small cells}

\item{height}{default height of normal antennas}

\item{height_small}{default height of small cells}

\item{tilt}{default (horizontal) tilt. Only applicable for directional antennas}

\item{beam_v}{default vertical beam width. Only applicable for directional antennas}

\item{beam_h}{default horizontal beam width. Only applicable for directional antennas}

\item{area_expension}{when \code{poly_shape} is \code{"pie"} or \code{"oval"}, the sizes of the polygons set to the area size of the corresponding Voronoi are multiplied by \code{area_expension}. These sizes are subsequently adjusted such that the range is bounded by \code{min_range} and \code{max_range} for normal cells and \code{min_range_small} and \code{max_range_small} for small cells. The purpose of \code{area_expension} is to allow overlap. This number (which should be greater than 1) contols the amout of overlap. The default value 1.5.}

\item{max_overlapping_cells}{maximum number of polygons that may overlap per raster cell. If the actual number exceeds this parameter, the \code{max_overlapping_cells} cells with the highest signal strength are selected}
}
\value{
parameter list
}
\description{
Set and update location model parameters. The function \code{location_model_parameters} specifies model parameters, which can be updated with \code{update_model_parameters}
}
