% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_cellplan.R
\name{rasterize_cellplan}
\alias{rasterize_cellplan}
\title{Rasterize cellplan}
\usage{
rasterize_cellplan(cp, cp_poly, raster, elevation, param)
}
\arguments{
\item{cp}{cellplan}

\item{cp_poly}{cellplan polygons}

\item{raster}{raster with indices}

\item{elevation}{raster with elevation data}

\item{param}{list}
}
\description{
Rasterize cellplan
}
