% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_location_model.R
\name{setup_location_model}
\alias{setup_location_model}
\title{Interactive cell modelling tool}
\usage{
setup_location_model(param, plot.height = 800)
}
\arguments{
\item{param}{list of parameters}

\item{plot.height}{heigth of the plots in the tool (in pixels)}
}
\description{
Tool to model the perceived signal strength
}
