% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_prop_model.R
\name{setup_prop_model}
\alias{setup_prop_model}
\title{Interactive tool to setup the propagation model}
\usage{
setup_prop_model(param, plot.height = 800)
}
\arguments{
\item{param}{list of parameters}

\item{plot.height}{heigth of the plots in the tool (in pixels)}
}
\description{
Tool to setup the propagation model parameters
}
