% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sig_strength_model.R
\name{setup_sig_strength_model}
\alias{setup_sig_strength_model}
\title{Interactive tool to setup the propagation model}
\usage{
setup_sig_strength_model(param = mobloc_param(), plot.height = 800)
}
\arguments{
\item{param}{parameter list created with \code{mobloc_param}}

\item{plot.height}{heigth of the plots in the tool (in pixels)}
}
\description{
Tool to setup the propagation model parameters
}
\examples{
# explore parameters
\dontrun{
    setup_sig_strength_model()
}
}
\seealso{
\href{../doc/mobloc.html}{\code{vignette("mobloc")}}
}
