% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_cells.R
\name{shiny_cells}
\alias{shiny_cells}
\title{Visualize signal strengths and probabilities per raster cell}
\usage{
shiny_cells(cp, cp_poly, raster, prob, param)
}
\arguments{
\item{cp}{cellplan}

\item{cp_poly}{cellplan polygons}

\item{raster}{raster}

\item{prob}{probabilities per raster cell}

\item{param}{parameter list}
}
\description{
Visualize signal strengths and probabilities per raster cell
}
