% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_strength.R
\name{signal_strength}
\alias{signal_strength}
\title{Signal strength}
\usage{
signal_strength(
  cx,
  cy,
  cz,
  direction,
  tilt,
  beam_h,
  beam_v,
  W,
  co,
  ple,
  param,
  enable = c("d", "h", "v")
)
}
\arguments{
\item{cx}{The x-coordinate of the source}

\item{cy}{The y-coordinate of the source}

\item{cz}{The z-coordinate of the source}

\item{direction}{The propagation direction}

\item{tilt}{The propagation tilt}

\item{beam_h}{The propagation horizontal beam width}

\item{beam_v}{The propagation vertical beam width}

\item{W}{The power (Watt) of the source}

\item{co}{The coordinates and id numbers of the tiles for which the signal strength should be calculated. Should be a data.frame with the variables \code{x}, \code{y}, \code{z}, \code{rid}}

\item{ple}{Path loss exponent}

\item{param}{mobloc parameters}

\item{enable}{character vector that determines which components are enabled for the calculation of signal strength. More specifically, these components determine whether to include signal loss as a function of distance (\code{"d"}), horizontal deviation (\code{"h"}) and vertical deviation (\code{"v"}) of the main direction.}
}
\description{
Low-level function to calculate signal strength. Used by compute_sig_strength.
}
