% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cluster.R
\name{start_cluster}
\alias{start_cluster}
\title{Setup cluster with desired number of cores.}
\usage{
start_cluster(nodes = NA)
}
\arguments{
\item{nodes}{Number of nodes that are used in this cluster. If missing the number of codes is used.}
}
\description{
Setup a cluster, only if not already defined.
}
