% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_posterior.R
\name{update_posterior_TA}
\alias{update_posterior_TA}
\title{Update posterior with Timing Advance}
\usage{
update_posterior_TA(post, raster, cp, param, elev = NULL)
}
\arguments{
\item{post}{posterior created with \code{\link{calculate_posterior}}}

\item{raster}{raster object that contains the raster tile index numbers (e.g. created with \code{\link{create_raster}})}

\item{cp}{cellplan}

\item{param}{parameter list created with \code{mobloc_param}}

\item{elev}{elevation data. If specified, it is used to calculate the distance using the z coordinate as well.}
}
\description{
Update posterior with Timing Advance. The posterior created with \code{\link{calculate_posterior}} will be used as a prior. The likelihood will be 1 if the grid call is contained in a specific TA band and 0 otherwise. When the parameter \code{TA_buffer > 0}, the TA band will be made broader. See the documentation of \code{\link{mobloc_param}} for details
}
\seealso{
\code{\link{calculate_posterior}} for examples
}
