% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{ZL_elevation}
\alias{ZL_elevation}
\title{Elevation data of Zuid-Limburg}
\usage{
data(ZL_elevation)
}
\description{
Elevation data of Zuid-Limburg, a province (NUTS 2 region) of the Netherlands.. It is a raster object of 100 x 100 meter grid tiles
}
