% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{ZL_landuse}
\alias{ZL_landuse}
\title{Landuse of Zuid-Limburg}
\usage{
data(ZL_landuse)
}
\description{
Landuse grouped of Zuid-Limburg, a province (NUTS 2 region) of the Netherlands. The landuse categories are groupd by: building, residential, forest, water and roads. Source: OpenStreetMap
}
