% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{ZL_resi}
\alias{ZL_resi}
\title{Residential areas of Zuid-Limburg}
\usage{
data(ZL_resi)
}
\description{
Residential of Zuid-Limburg, raster object of 100 x 100 meter cells, where for each cell the fraction of area is specified which is tagged as residential. Source: OpenStreetMap
}
