% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{calculate_mobloc}
\alias{calculate_mobloc}
\title{Calculate mobile location}
\usage{
calculate_mobloc(prop, prior)
}
\arguments{
\item{prop}{propagation object, the result of \code{\link{process_cellplan}}}

\item{prior}{prior object, the result of \code{\link{extract_uniform_prior}}, \code{\link{extract_prior}}, or \code{\link{extract_network_prior}}}
}
\description{
Calculate the mobile location given the calculated propagation and a prior.
}
