% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cellplan.R
\name{check_cellplan}
\alias{check_cellplan}
\title{Check cellplan}
\usage{
check_cellplan(cp, param, land = NULL, elevation = NULL)
}
\arguments{
\item{cp}{cellplan \code{sf} object containing the antenna data. Each data record should be a point (i.e., `st_geometry_type(cp)` should return `POINT`s). The variables (of which only the first is required) are used:
\itemize{
\item \code{height}. Height of the antenna. If omitted, the default value \code{height} from the parameter list \code{param} will be used.
\item \code{z}. Required unless \code{elevation} is specified. Note that \code{z = elevation + height}. So if the arugment \code{elevation} is specified, \code{z} will automatically be derived.
\item \code{direction}. Direction of the antanna in degrees. Use \code{NA} for omnidirectional antennas.
\item \code{tilt}. Tilt of the antennas in degrees. Only applicable for directional cells. If omitted, the default value \code{tilt} from the parameter list \code{param} will be used.
\item \code{beam_h}. Horizontal beam width in degrees. The signal loss at \code{-beam_h/2} and \code{+beam_h/2} degrees is 3 dB. Run \code{radiation_plot(beam_width = 65, db_back = -30)}. If omitted, the default value \code{beam_h} from the parameter list \code{param} will be used.
\item \code{beam_v}. Vertical beam width in degrees. The signal loss at \code{-beam_v/2} and \code{+beam_v/2} degrees is 3 dB. Run \code{radiation_plot(type = "e", beam_width = 9, db_back = -30)}. If omitted, the default value \code{beam_v} from the parameter list \code{param} will be used.
\item \code{small}. Logical value that determines whether the antenna is a 'small cell'. If omitted, it will be set to \code{FALSE}. In the \code{mobloc} package, small cells have different default values for a couple of parameters (i.e. the \code{"_small"} parameters in \code{param}).
\item \code{range}. The maximum range of the antenna. If omitted, the value \code{max_range} from the parameter list \code{param} will be used. If \code{small} is defined, the value \code{max_range_small} is used for each antenna for which \code{small == TRUE}.
}}

\item{param}{parameter list. See \code{\link{location_model_parameters}}.}

\item{land}{land polygon. If specifies, it checks if all antennas are contained inside it}

\item{elevation}{see argument \code{cp} (variable \code{z})}
}
\description{
Function to the check cellplan. It checks the format and whether are all required variables present. When \code{land}, a mutlipolygon object that defines the region of interest, is specified, it checks whether all antennas are inside this region.
}
