% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{combine_raster_layers}
\alias{combine_raster_layers}
\title{Combine raster layers to a single raster layer}
\usage{
combine_raster_layers(..., weights = NULL)
}
\arguments{
\item{...}{Either one raster brick/stack, or multiple raster layer objects}

\item{weights}{numeric vector that specifies the weights of the raster layers}
}
\description{
Combine raster layers to a single raster layer. Weights can be assigned to the raster layers.
}
