% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cellplan_polygons.R
\name{create_cellplan_polygons}
\alias{create_cellplan_polygons}
\title{Create cellplan polygons}
\usage{
create_cellplan_polygons(cp, land, bbox, param)
}
\arguments{
\item{cp}{cellplan}

\item{land}{land}

\item{bbox}{bounding box}

\item{param}{parameters}
}
\value{
cellplan polygons
}
\description{
Create cellplan polygons
}
