% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coverage_map.R
\name{create_coverage_map}
\alias{create_coverage_map}
\alias{create_best_server_map}
\title{Functions to estimate a coverage map and a best server map.}
\usage{
create_coverage_map(prop, raster, type = c("dBm", "s"),
  antennas = NULL)

create_best_server_map(prop, raster, antennas = NULL)
}
\arguments{
\item{prop}{propagation model}

\item{raster}{raster raster object}

\item{type}{either \code{"dBm"} for absolute signal strength values and \code{"s"} for signal quality values}

\item{antennas}{selection of antennas}
}
\description{
The function \code{create_coverage_map} is used to create a coverage map, and the function \code{create_best_server_map} to create a best (area) server map. The former shows the best signal strength per raster cell (either in dBm or in signal quality). The latter shows the best antenna per raster cell.
}
