% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{create_prior}
\alias{create_prior}
\alias{create_uniform_prior}
\alias{create_network_prior}
\alias{prior_filter}
\title{Create a prior distribution}
\usage{
create_prior(..., name = "composite", weights = NULL)

create_uniform_prior(raster)

create_network_prior(prop, raster)

prior_filter(prior, region)
}
\arguments{
\item{...}{one of more raster objects, or a list of it. When multiple raster layers are used, these are first combined using a weighed sum (see argument \code{weights} and the underlying function \code{\link{combine_raster_layers}}). After that, these values are normalised such that the values add up to 1}

\item{name}{name of the prior}

\item{weights}{The weights of the raster objects, which should be a vector of the same length as the number of specified raster layers. See also the underlying function \code{\link{combine_raster_layers}}.}

\item{raster}{raster object that contains the raster tile index numbers (e.g. created with \code{\link{create_raster}})}

\item{prop}{a propagation object, which is the result of \code{\link{process_cellplan}}}

\item{prior}{object created with \code{create_prior}, \code{create_network_prior} or \code{create_uniform_prior}}

\item{region}{a polygons object (class \code{sf}) that specifies a region. If the object contains multiple polygons, they are considered as one.}
}
\value{
data frame with two columns, raster id number \code{rid} and probability \code{p}. These probabilities will add up to 1.
}
\description{
Functions to create a prior distribution. A prior distribution is a raster layer where all values add up to 1. The function \code{create_prior} creates a prior using one or more raster layers. The function \code{create_uniform_prior} creates a uniform prior. The function \code{create_network_prior} extracts a prior from the propagation model. This is based on the assumption that an MNO will place antennas where they expect people. The function \code{prior_filter} can be used to filter a prior raster using a polygons object (for instance, the administrative region of interest).
}
\examples{
\dontrun{
    # set parameters
    ZL_param <- prop_param()

    # load data
    data("ZL_cellplan", "ZL_muni", "ZL_elevation", "ZL_landuse")

    # create environment layer (needed to calculate path loss exponent (ple))
    ZL_envir <- combine_raster_layers(ZL_landuse, weights = c(1, 1, 1, 0, 0))

    # validate cellplan
    ZL_cellplan <- validate_cellplan(ZL_cellplan, param = ZL_param, region = ZL_muni, envir = ZL_envir, elevation = ZL_elevation)

    # create raster
    ZL_bbox <- sf::st_bbox(c(xmin = 4012000, ymin = 3077000, xmax = 4048000, ymax = 3117000), crs = st_crs(3035))
    ZL_raster <- create_raster(ZL_bbox)

    # process cellplan (result is propagation model and connection likelihood)
    ZL_prop <- process_cellplan(cp = ZL_cellplan, raster = ZL_raster, elevation = ZL_elevation, param = ZL_param)

    # create priors
    ZL_uniform_prior <- create_uniform_prior(ZL_raster)
    ZL_network_prior <- create_network_prior(ZL_prop, ZL_raster)
    ZL_landuse_prior <- create_prior(ZL_landuse, weights = c(1, 1, .1, 0, .5))
    ZL_comp_prior <- create_prior(ZL_network_prior, ZL_landuse_prior, weights = c(.25, .75))
}
}
\seealso{
\href{../doc/mobloc.html}{\code{vignette("mobloc")}}
}
