% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_rasterize.R
\name{create_raster}
\alias{create_raster}
\title{Create empty raster}
\usage{
create_raster(x, tile.size = 100)
}
\arguments{
\item{x}{either a bounding box (see \code{\link[sf:st_bbox]{st_bbox}}) where the projection (\code{crs}) must be specified, or a raster object (where the elevation object is supposed to be used)}

\item{tile.size}{cell size (which is both the width and the height) in meters (assuming the coordinate system is specified in meters)}
}
\value{
raster layer
}
\description{
Create empty raster from a bounding box with a specified projection.
}
\examples{
# create raster
ZL_bbox <- sf::st_bbox(c(xmin = 4012000, ymin = 3077000, xmax = 4048000, ymax = 3117000), crs = sf::st_crs(3035))
ZL_raster <- create_raster(ZL_bbox)
}
