% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mobloc.R
\name{explore_mobloc}
\alias{explore_mobloc}
\title{Explore the Visualize signal strengths and probabilities per raster cell}
\usage{
explore_mobloc(cp, raster, prop, priorlist = NULL, param)
}
\arguments{
\item{cp}{cellplan}

\item{raster}{raster}

\item{prop}{propagation model}

\item{priorlist}{list of priors}

\item{param}{parameter list}
}
\description{
Visualize signal strengths and probabilities per raster cell
}
