% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mobloc.R
\name{explore_mobloc}
\alias{explore_mobloc}
\title{Explore the Visualize signal strengths and probabilities per raster cell}
\usage{
explore_mobloc(cp, cp_poly, raster, prop, priorlist = NULL, param,
  tm = NULL)
}
\arguments{
\item{cp}{cellplan}

\item{cp_poly}{cellplan polygons}

\item{raster}{raster}

\item{prop}{propagation model}

\item{priorlist}{list of priors}

\item{param}{parameter list}

\item{tm}{tmap object (optional), that is plotted on top of the map}
}
\description{
Visualize signal strengths and probabilities per raster cell
}
