% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{extract_prior}
\alias{extract_prior}
\alias{extract_network_prior}
\alias{extract_uniform_prior}
\title{Extract prior information}
\usage{
extract_prior(x, prop)

extract_network_prior(prop)

extract_uniform_prior(prop)
}
\arguments{
\item{x}{a raster object with per raster cell a prior value.}

\item{prop}{a propagation object, which is the result of \code{\link{process_cellplan}}}
}
\value{
data frame with two columns, raster id number \code{rid} and probability \code{p}. These probabilities will add up to 1.
}
\description{
Extract prior information. The function \code{extract_prior} extracts the prior from a raster object. The function \code{extract_uniform_prior} extracts a uniform prior from the propagation model. The function \code{extract_network_prior} extracts a prior from the propagation model.
}
