% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cellplan.R
\name{process_cellplan}
\alias{process_cellplan}
\title{Rasterize cellplan}
\usage{
process_cellplan(cp, raster, elevation, param, region = NULL)
}
\arguments{
\item{cp}{cellplan, validated with \code{\link{validate_cellplan}}}

\item{raster}{raster object that contains the raster tile index numbers (e.g. created with \code{\link{create_raster}})}

\item{elevation}{raster with elevation data}

\item{param}{parameter list created with \code{prop_param}}

\item{region}{polygon shape. If specified, only the signal strength will be calculated for raster tiles inside the polygons}
}
\description{
Rasterize cellplan
}
