% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_param.R
\name{prop_param}
\alias{prop_param}
\alias{update_prop_param}
\title{Set and update propagation model parameters}
\usage{
prop_param(W = 10, W_small = 5, ple = 3.25, ple_small = 6,
  ple_0 = 2.5, ple_1 = 4, azim_min3dB = 65, azim_dB_back = -30,
  elev_min3dB = 9, elev_dB_back = -30, dBm_mid = -92.5,
  dBm_width = 5, range = 20000, range_small = 5000, height = 30,
  height_small = 8, tilt = 5, beam_v = 9, beam_h = 65,
  dBm_th = -120, max_overlapping_cells = 100)

update_prop_param(param, ...)
}
\arguments{
\item{W}{default power in Watt of a normal antennas (placed in a cell tower or rooftop site)}

\item{W_small}{default power in Watt of a small cell (omnidirectional)}

\item{ple}{default path loss exponent}

\item{ple_small}{path loss exponent for small cells}

\item{ple_0}{path loss exponent for free space}

\item{ple_1}{path loss exponent for dense environments}

\item{azim_min3dB}{default horizontal beam width. At \code{azim_min3dB/2}, the signal strength is halved (so -3dBM)}

\item{azim_dB_back}{difference in signal strength between front and back}

\item{elev_min3dB}{default vertical beam width. At \code{elev_min3dB/2}, the signal strength is halved (so -3dBM)}

\item{elev_dB_back}{difference in signal strength between front and back}

\item{dBm_mid}{middle point in the logistic function to map signal strength to probability}

\item{dBm_width}{width of the logistic function to map signal strength to probability}

\item{range}{maximum range of normal antennas}

\item{range_small}{maximum range of small cells}

\item{height}{default height of normal antennas}

\item{height_small}{default height of small cells}

\item{tilt}{default (horizontal) tilt. Only applicable for directional antennas}

\item{beam_v}{default vertical beam width. Only applicable for directional antennas}

\item{beam_h}{default horizontal beam width. Only applicable for directional antennas}

\item{dBm_th}{dBm threshold}

\item{max_overlapping_cells}{maximum number of polygons that may overlap per raster cell. If the actual number exceeds this parameter, the \code{max_overlapping_cells} cells with the highest signal strength are selected}

\item{param}{parameter list created with \code{prop_param}}

\item{...}{parameter updates, see arguments of \code{prop_param}}
}
\value{
parameter list
}
\description{
Set and update propagation model parameters. The function \code{prop_param} specifies model parameters, which can be updated with \code{update_prop_param}
}
