% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_param.R
\name{prop_param}
\alias{prop_param}
\title{Set the propagation model parameters}
\usage{
prop_param(W = 10, W_small = 5, ple = 3.75, ple_small = 6,
  ple_0 = 3.5, ple_1 = 4, dBm_mid = -92.5, dBm_width = 5,
  midpoint = -92.5, steepness = 1/5, range = 10000,
  range_small = 1000, height = 30, height_small = 8, tilt = 5,
  beam_v = 9, beam_h = 65, azim_dB_back = -30, elev_dB_back = -30,
  sig_q_th = 0.005, max_overlapping_antennas = 100, TA_step = 78.12,
  TA_max = 1282, TA_buffer = 1)
}
\arguments{
\item{W}{default power in Watt of a normal antennas (placed in a cell tower or rooftop site)}

\item{W_small}{default power in Watt of a small cell (omnidirectional)}

\item{ple}{default path loss exponent}

\item{ple_small}{path loss exponent for small cells}

\item{ple_0}{path loss exponent for free space}

\item{ple_1}{path loss exponent for dense environments}

\item{dBm_mid}{middle point in the logistic function to map signal strength to probability}

\item{dBm_width}{width of the logistic function to map signal strength to probability}

\item{midpoint}{midpoint of the logistic function used to map signal strength to signal quality}

\item{steepness}{steepness of the logistic function used to map signal strength to signal quality}

\item{range}{maximum range of normal antennas}

\item{range_small}{maximum range of small cells}

\item{height}{default height of normal antennas}

\item{height_small}{default height of small cells}

\item{tilt}{default (horizontal) tilt. Only applicable for directional antennas}

\item{beam_v}{default vertical beam width. Only applicable for directional antennas}

\item{beam_h}{default horizontal beam width. Only applicable for directional antennas}

\item{azim_dB_back}{difference in signal strength between front and back}

\item{elev_dB_back}{difference in signal strength between front and back}

\item{sig_q_th}{signal quality threshold}

\item{max_overlapping_antennas}{maximum number of antennas that may overlap per raster tile. If the actual number exceeds this parameter, the \code{max_overlapping_antennas} cells with the highest signal strength are selected}

\item{TA_step}{meters that correspond to one Timing Advance (TA) step. This parameter depends on the network technology and psychical properties such as air pressure. In GSM networks it is approximately 554 meters \url{https://people.csail.mit.edu/bkph/cellular_repeater_TA.shtml}, and LTE (4G) networks 78.12 meters.\url{https://www.scribd.com/doc/290553975/Timing-Advance-TA-in-LTE}}

\item{TA_max}{maximum Timing Advance (TA) value (integer). In other words, TA can have a value between 0 and \code{TA_max}. In GSM it is 63, and in LTE 1282.}

\item{TA_buffer}{buffer to prevent artifacts in the TA to grid tile conversion. These artifacts occur when \code{TA_step} is similar or smaller than the width of a grid tile. \code{TA_buffer} is an integer that determines the number of TA steps that are added in front of behind the actual TA band}
}
\value{
parameter list
}
\description{
Set the propagation model parameters. The result is a list.
}
\examples{
# set parameters
ZL_param <- prop_param()
}
\seealso{
\href{../doc/mobloc.html}{\code{vignette("mobloc")}}
}
